requirements.txt
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
gTTS>=2.3.2
python-dotenv>=1.0.0
